8268821: Split systemDictionaryShared.cpp
