+++ src/hotspot/share/classfile/systemDictionaryShared.cpp
@@ -29,6 +29,7 @@
 #include "cds/dynamicArchive.hpp"
 #include "cds/filemap.hpp"
 #include "cds/heapShared.hpp"
+#include "cds/cdsProtectionDomain.hpp"
 #include "cds/metaspaceShared.hpp"
 #include "classfile/classFileStream.hpp"
 #include "classfile/classLoader.hpp"
@@ -60,7 +61,6 @@
 #include "oops/oop.inline.hpp"
 #include "oops/oopHandle.inline.hpp"
 #include "oops/typeArrayOop.inline.hpp"
-#include "prims/jvmtiExport.hpp"
 #include "runtime/arguments.hpp"
 #include "runtime/handles.inline.hpp"
 #include "runtime/java.hpp"
@@ -70,870 +70,271 @@
 #include "utilities/resourceHash.hpp"
 #include "utilities/stringUtils.hpp"
 
+DumpTimeSharedClassTable* SystemDictionaryShared::_dumptime_table = NULL;
+DumpTimeLambdaProxyClassDictionary* SystemDictionaryShared::_dumptime_lambda_proxy_class_dictionary = NULL;
+// SystemDictionaries in the base layer static archive
+RunTimeSharedDictionary SystemDictionaryShared::_builtin_dictionary;
+RunTimeSharedDictionary SystemDictionaryShared::_unregistered_dictionary;
+// SystemDictionaries in the top layer dynamic archive
+RunTimeSharedDictionary SystemDictionaryShared::_dynamic_builtin_dictionary;
+RunTimeSharedDictionary SystemDictionaryShared::_dynamic_unregistered_dictionary;
 
-OopHandle SystemDictionaryShared::_shared_protection_domains;
-OopHandle SystemDictionaryShared::_shared_jar_urls;
-OopHandle SystemDictionaryShared::_shared_jar_manifests;
+LambdaProxyClassDictionary SystemDictionaryShared::_lambda_proxy_class_dictionary;
+LambdaProxyClassDictionary SystemDictionaryShared::_dynamic_lambda_proxy_class_dictionary;
+
 DEBUG_ONLY(bool SystemDictionaryShared::_no_class_loading_should_happen = false;)
 bool SystemDictionaryShared::_dump_in_progress = false;
 
-class DumpTimeSharedClassInfo: public CHeapObj<mtClass> {
-  bool                         _excluded;
-  bool                         _is_early_klass;
-  bool                         _has_checked_exclusion;
-public:
-  struct DTLoaderConstraint {
-    Symbol* _name;
-    char _loader_type1;
-    char _loader_type2;
-    DTLoaderConstraint(Symbol* name, char l1, char l2) : _name(name), _loader_type1(l1), _loader_type2(l2) {
-      _name->increment_refcount();
-    }
-    DTLoaderConstraint() : _name(NULL), _loader_type1('0'), _loader_type2('0') {}
-    bool equals(const DTLoaderConstraint& t) {
-      return t._name == _name &&
-             ((t._loader_type1 == _loader_type1 && t._loader_type2 == _loader_type2) ||
-              (t._loader_type2 == _loader_type1 && t._loader_type1 == _loader_type2));
-    }
-  };
+InstanceKlass* SystemDictionaryShared::load_shared_class_for_builtin_loader(
+                 Symbol* class_name, Handle class_loader, TRAPS) {
+  assert(UseSharedSpaces, "must be");
+  InstanceKlass* ik = find_builtin_class(class_name);
 
-  struct DTVerifierConstraint {
-    Symbol* _name;
-    Symbol* _from_name;
-    DTVerifierConstraint() : _name(NULL), _from_name(NULL) {}
-    DTVerifierConstraint(Symbol* n, Symbol* fn) : _name(n), _from_name(fn) {
-      _name->increment_refcount();
-      _from_name->increment_refcount();
-    }
-  };
-
-  InstanceKlass*               _klass;
-  InstanceKlass*               _nest_host;
-  bool                         _failed_verification;
-  bool                         _is_archived_lambda_proxy;
-  int                          _id;
-  int                          _clsfile_size;
-  int                          _clsfile_crc32;
-  GrowableArray<DTVerifierConstraint>* _verifier_constraints;
-  GrowableArray<char>*                 _verifier_constraint_flags;
-  GrowableArray<DTLoaderConstraint>* _loader_constraints;
-
-  DumpTimeSharedClassInfo() {
-    _klass = NULL;
-    _nest_host = NULL;
-    _failed_verification = false;
-    _is_archived_lambda_proxy = false;
-    _has_checked_exclusion = false;
-    _id = -1;
-    _clsfile_size = -1;
-    _clsfile_crc32 = -1;
-    _excluded = false;
-    _is_early_klass = JvmtiExport::is_early_phase();
-    _verifier_constraints = NULL;
-    _verifier_constraint_flags = NULL;
-    _loader_constraints = NULL;
-  }
-
-  void add_verification_constraint(InstanceKlass* k, Symbol* name,
-         Symbol* from_name, bool from_field_is_protected, bool from_is_array, bool from_is_object);
-  void record_linking_constraint(Symbol* name, Handle loader1, Handle loader2);
-
-  bool is_builtin() {
-    return SystemDictionaryShared::is_builtin(_klass);
-  }
-
-  int num_verifier_constraints() {
-    if (_verifier_constraint_flags != NULL) {
-      return _verifier_constraint_flags->length();
-    } else {
-      return 0;
+  if (ik != NULL && !ik->shared_loading_failed()) {
+    if ((SystemDictionary::is_system_class_loader(class_loader()) && ik->is_shared_app_class())  ||
+        (SystemDictionary::is_platform_class_loader(class_loader()) && ik->is_shared_platform_class())) {
+      SharedClassLoadingMark slm(THREAD, ik);
+      PackageEntry* pkg_entry = CDSProtectionDomain::get_package_entry_from_class(ik, class_loader);
+      Handle protection_domain =
+        CDSProtectionDomain::init_security_info(class_loader, ik, pkg_entry, CHECK_NULL);
+      return load_shared_class(ik, class_loader, protection_domain, NULL, pkg_entry, THREAD);
     }
   }
-
-  int num_loader_constraints() {
-    if (_loader_constraints != NULL) {
-      return _loader_constraints->length();
-    } else {
-      return 0;
-    }
-  }
-
-  void metaspace_pointers_do(MetaspaceClosure* it) {
-    it->push(&_klass);
-    it->push(&_nest_host);
-    if (_verifier_constraints != NULL) {
-      for (int i = 0; i < _verifier_constraints->length(); i++) {
-        DTVerifierConstraint* cons = _verifier_constraints->adr_at(i);
-        it->push(&cons->_name);
-        it->push(&cons->_from_name);
-      }
-    }
-    if (_loader_constraints != NULL) {
-      for (int i = 0; i < _loader_constraints->length(); i++) {
-        DTLoaderConstraint* lc = _loader_constraints->adr_at(i);
-        it->push(&lc->_name);
-      }
-    }
-  }
-
-  bool is_excluded() {
-    // _klass may become NULL due to DynamicArchiveBuilder::set_to_null
-    return _excluded || _failed_verification || _klass == NULL;
-  }
-
-  // Was this class loaded while JvmtiExport::is_early_phase()==true
-  bool is_early_klass() {
-    return _is_early_klass;
-  }
-
-  // simple accessors
-  void set_excluded()                               { _excluded = true; }
-  bool has_checked_exclusion() const                { return _has_checked_exclusion; }
-  void set_has_checked_exclusion()                  { _has_checked_exclusion = true; }
-  bool failed_verification() const                  { return _failed_verification; }
-  void set_failed_verification()                    { _failed_verification = true; }
-  InstanceKlass* nest_host() const                  { return _nest_host; }
-  void set_nest_host(InstanceKlass* nest_host)      { _nest_host = nest_host; }
-};
-
-inline unsigned DumpTimeSharedClassTable_hash(InstanceKlass* const& k) {
-  if (DumpSharedSpaces) {
-    // Deterministic archive contents
-    uintx delta = k->name() - MetaspaceShared::symbol_rs_base();
-    return primitive_hash<uintx>(delta);
-  } else {
-    // Deterministic archive is not possible because classes can be loaded
-    // in multiple threads.
-    return primitive_hash<InstanceKlass*>(k);
-  }
+  return NULL;
 }
 
-class DumpTimeSharedClassTable: public ResourceHashtable<
-  InstanceKlass*,
-  DumpTimeSharedClassInfo,
-  &DumpTimeSharedClassTable_hash,
-  primitive_equals<InstanceKlass*>,
-  15889, // prime number
-  ResourceObj::C_HEAP>
-{
-  int _builtin_count;
-  int _unregistered_count;
-public:
-  DumpTimeSharedClassInfo* find_or_allocate_info_for(InstanceKlass* k, bool dump_in_progress) {
-    bool created = false;
-    DumpTimeSharedClassInfo* p;
-    if (!dump_in_progress) {
-      p = put_if_absent(k, &created);
-    } else {
-      p = get(k);
+// This function is called for loading only UNREGISTERED classes
+InstanceKlass* SystemDictionaryShared::lookup_from_stream(Symbol* class_name,
+                                                          Handle class_loader,
+                                                          Handle protection_domain,
+                                                          const ClassFileStream* cfs,
+                                                          TRAPS) {
+  if (!UseSharedSpaces) {
+    return NULL;
+  }
+  if (class_name == NULL) {  // don't do this for hidden classes
+    return NULL;
+  }
+  if (class_loader.is_null() ||
+      SystemDictionary::is_system_class_loader(class_loader()) ||
+      SystemDictionary::is_platform_class_loader(class_loader())) {
+    // Do nothing for the BUILTIN loaders.
+    return NULL;
+  }
+
+  const RunTimeClassInfo* record = find_record(&_unregistered_dictionary, &_dynamic_unregistered_dictionary, class_name);
+  if (record == NULL) {
+    return NULL;
+  }
+
+  int clsfile_size  = cfs->length();
+  int clsfile_crc32 = ClassLoader::crc32(0, (const char*)cfs->buffer(), cfs->length());
+
+  if (!record->matches(clsfile_size, clsfile_crc32)) {
+    return NULL;
+  }
+
+  return acquire_class_for_current_thread(record->_klass, class_loader,
+                                          protection_domain, cfs,
+                                          THREAD);
+}
+
+InstanceKlass* SystemDictionaryShared::acquire_class_for_current_thread(
+                   InstanceKlass *ik,
+                   Handle class_loader,
+                   Handle protection_domain,
+                   const ClassFileStream *cfs,
+                   TRAPS) {
+  ClassLoaderData* loader_data = ClassLoaderData::class_loader_data(class_loader());
+
+  {
+    MutexLocker mu(THREAD, SharedDictionary_lock);
+    if (ik->class_loader_data() != NULL) {
+      //    ik is already loaded (by this loader or by a different loader)
+      // or ik is being loaded by a different thread (by this loader or by a different loader)
+      return NULL;
     }
-    if (created) {
-      assert(!SystemDictionaryShared::no_class_loading_should_happen(),
-             "no new classes can be loaded while dumping archive");
-      p->_klass = k;
-    } else {
-      if (!dump_in_progress) {
-        assert(p->_klass == k, "Sanity");
-      }
-    }
-    return p;
+
+    // No other thread has acquired this yet, so give it to *this thread*
+    ik->set_class_loader_data(loader_data);
   }
 
-  class CountClassByCategory : StackObj {
-    DumpTimeSharedClassTable* _table;
-  public:
-    CountClassByCategory(DumpTimeSharedClassTable* table) : _table(table) {}
-    bool do_entry(InstanceKlass* k, DumpTimeSharedClassInfo& info) {
-      if (!info.is_excluded()) {
-        if (info.is_builtin()) {
-          ++ _table->_builtin_count;
-        } else {
-          ++ _table->_unregistered_count;
-        }
-      }
-      return true; // keep on iterating
-    }
-  };
+  // No longer holding SharedDictionary_lock
+  // No need to lock, as <ik> can be held only by a single thread.
+  loader_data->add_class(ik);
 
-  void update_counts() {
-    _builtin_count = 0;
-    _unregistered_count = 0;
-    CountClassByCategory counter(this);
-    iterate(&counter);
+  // Get the package entry.
+  PackageEntry* pkg_entry = CDSProtectionDomain::get_package_entry_from_class(ik, class_loader);
+
+  // Load and check super/interfaces, restore unsharable info
+  InstanceKlass* shared_klass = load_shared_class(ik, class_loader, protection_domain,
+                                                  cfs, pkg_entry, THREAD);
+  if (shared_klass == NULL || HAS_PENDING_EXCEPTION) {
+    // TODO: clean up <ik> so it can be used again
+    return NULL;
   }
 
-  int count_of(bool is_builtin) const {
-    if (is_builtin) {
-      return _builtin_count;
-    } else {
-      return _unregistered_count;
-    }
-  }
-};
+  return shared_klass;
+}
 
-class LambdaProxyClassKey {
-  InstanceKlass* _caller_ik;
-  Symbol*        _invoked_name;
-  Symbol*        _invoked_type;
-  Symbol*        _method_type;
-  Method*        _member_method;
-  Symbol*        _instantiated_method_type;
+void SystemDictionaryShared::start_dumping() {
+  MutexLocker ml(DumpTimeTable_lock, Mutex::_no_safepoint_check_flag);
+  _dump_in_progress = true;
+}
 
-public:
-  LambdaProxyClassKey(InstanceKlass* caller_ik,
-                      Symbol*        invoked_name,
-                      Symbol*        invoked_type,
-                      Symbol*        method_type,
-                      Method*        member_method,
-                      Symbol*        instantiated_method_type) :
-    _caller_ik(caller_ik),
-    _invoked_name(invoked_name),
-    _invoked_type(invoked_type),
-    _method_type(method_type),
-    _member_method(member_method),
-    _instantiated_method_type(instantiated_method_type) {}
+DumpTimeClassInfo* SystemDictionaryShared::find_or_allocate_info_for(InstanceKlass* k) {
+  MutexLocker ml(DumpTimeTable_lock, Mutex::_no_safepoint_check_flag);
+  return find_or_allocate_info_for_locked(k);
+}
 
-  void metaspace_pointers_do(MetaspaceClosure* it) {
-    it->push(&_caller_ik);
-    it->push(&_invoked_name);
-    it->push(&_invoked_type);
-    it->push(&_method_type);
-    it->push(&_member_method);
-    it->push(&_instantiated_method_type);
-  }
-
-  void mark_pointers() {
-    ArchivePtrMarker::mark_pointer(&_caller_ik);
-    ArchivePtrMarker::mark_pointer(&_instantiated_method_type);
-    ArchivePtrMarker::mark_pointer(&_invoked_name);
-    ArchivePtrMarker::mark_pointer(&_invoked_type);
-    ArchivePtrMarker::mark_pointer(&_member_method);
-    ArchivePtrMarker::mark_pointer(&_method_type);
-  }
-
-  bool equals(LambdaProxyClassKey const& other) const {
-    return _caller_ik == other._caller_ik &&
-           _invoked_name == other._invoked_name &&
-           _invoked_type == other._invoked_type &&
-           _method_type == other._method_type &&
-           _member_method == other._member_method &&
-           _instantiated_method_type == other._instantiated_method_type;
-  }
-
-  unsigned int hash() const {
-    return SystemDictionaryShared::hash_for_shared_dictionary((address)_caller_ik) +
-           SystemDictionaryShared::hash_for_shared_dictionary((address)_invoked_name) +
-           SystemDictionaryShared::hash_for_shared_dictionary((address)_invoked_type) +
-           SystemDictionaryShared::hash_for_shared_dictionary((address)_method_type) +
-           SystemDictionaryShared::hash_for_shared_dictionary((address)_instantiated_method_type);
-  }
-
-  static unsigned int dumptime_hash(Symbol* sym)  {
-    if (sym == NULL) {
-      // _invoked_name maybe NULL
-      return 0;
-    }
-    return java_lang_String::hash_code((const jbyte*)sym->bytes(), sym->utf8_length());
-  }
-
-  unsigned int dumptime_hash() const {
-    return dumptime_hash(_caller_ik->name()) +
-           dumptime_hash(_invoked_name) +
-           dumptime_hash(_invoked_type) +
-           dumptime_hash(_method_type) +
-           dumptime_hash(_instantiated_method_type);
-  }
-
-  static inline unsigned int DUMPTIME_HASH(LambdaProxyClassKey const& key) {
-    return (key.dumptime_hash());
-  }
-
-  static inline bool DUMPTIME_EQUALS(
-      LambdaProxyClassKey const& k1, LambdaProxyClassKey const& k2) {
-    return (k1.equals(k2));
-  }
-
-  InstanceKlass* caller_ik() const { return _caller_ik; }
-};
-
-
-class DumpTimeLambdaProxyClassInfo {
-public:
-  GrowableArray<InstanceKlass*>* _proxy_klasses;
-  DumpTimeLambdaProxyClassInfo() : _proxy_klasses(NULL) {}
-  void add_proxy_klass(InstanceKlass* proxy_klass) {
-    if (_proxy_klasses == NULL) {
-      _proxy_klasses = new (ResourceObj::C_HEAP, mtClassShared)GrowableArray<InstanceKlass*>(5, mtClassShared);
-    }
-    assert(_proxy_klasses != NULL, "sanity");
-    _proxy_klasses->append(proxy_klass);
-  }
-
-  void metaspace_pointers_do(MetaspaceClosure* it) {
-    for (int i=0; i<_proxy_klasses->length(); i++) {
-      it->push(_proxy_klasses->adr_at(i));
-    }
-  }
-};
-
-class RunTimeLambdaProxyClassInfo {
-  LambdaProxyClassKey _key;
-  InstanceKlass* _proxy_klass_head;
-public:
-  RunTimeLambdaProxyClassInfo(LambdaProxyClassKey key, InstanceKlass* proxy_klass_head) :
-    _key(key), _proxy_klass_head(proxy_klass_head) {}
-
-  InstanceKlass* proxy_klass_head() const { return _proxy_klass_head; }
-
-  // Used by LambdaProxyClassDictionary to implement OffsetCompactHashtable::EQUALS
-  static inline bool EQUALS(
-       const RunTimeLambdaProxyClassInfo* value, LambdaProxyClassKey* key, int len_unused) {
-    return (value->_key.equals(*key));
-  }
-  void init(LambdaProxyClassKey& key, DumpTimeLambdaProxyClassInfo& info) {
-    _key = key;
-    _key.mark_pointers();
-    _proxy_klass_head = info._proxy_klasses->at(0);
-    ArchivePtrMarker::mark_pointer(&_proxy_klass_head);
-  }
-
-  unsigned int hash() const {
-    return _key.hash();
-  }
-  LambdaProxyClassKey key() const {
-    return _key;
-  }
-};
-
-class LambdaProxyClassDictionary : public OffsetCompactHashtable<
-  LambdaProxyClassKey*,
-  const RunTimeLambdaProxyClassInfo*,
-  RunTimeLambdaProxyClassInfo::EQUALS> {};
-
-LambdaProxyClassDictionary _lambda_proxy_class_dictionary;
-
-LambdaProxyClassDictionary _dynamic_lambda_proxy_class_dictionary;
-
-class DumpTimeLambdaProxyClassDictionary
-  : public ResourceHashtable<LambdaProxyClassKey,
-                             DumpTimeLambdaProxyClassInfo,
-                             LambdaProxyClassKey::DUMPTIME_HASH,
-                             LambdaProxyClassKey::DUMPTIME_EQUALS,
-                             137, // prime number
-                             ResourceObj::C_HEAP> {
-public:
-  int _count;
-};
-
-DumpTimeLambdaProxyClassDictionary* _dumptime_lambda_proxy_class_dictionary = NULL;
-
-static void add_to_dump_time_lambda_proxy_class_dictionary(LambdaProxyClassKey key,
-                                                           InstanceKlass* proxy_klass) {
+DumpTimeClassInfo* SystemDictionaryShared::find_or_allocate_info_for_locked(InstanceKlass* k) {
   assert_lock_strong(DumpTimeTable_lock);
-  if (_dumptime_lambda_proxy_class_dictionary == NULL) {
-    _dumptime_lambda_proxy_class_dictionary =
-      new (ResourceObj::C_HEAP, mtClass)DumpTimeLambdaProxyClassDictionary();
+  if (_dumptime_table == NULL) {
+    _dumptime_table = new (ResourceObj::C_HEAP, mtClass)DumpTimeSharedClassTable();
   }
-  DumpTimeLambdaProxyClassInfo* lambda_info = _dumptime_lambda_proxy_class_dictionary->get(key);
-  if (lambda_info == NULL) {
-    DumpTimeLambdaProxyClassInfo info;
-    info.add_proxy_klass(proxy_klass);
-    _dumptime_lambda_proxy_class_dictionary->put(key, info);
-    //lambda_info = _dumptime_lambda_proxy_class_dictionary->get(key);
-    //assert(lambda_info->_proxy_klass == proxy_klass, "must be"); // debug only -- remove
-    ++_dumptime_lambda_proxy_class_dictionary->_count;
+  return _dumptime_table->find_or_allocate_info_for(k, _dump_in_progress);
+}
+
+bool SystemDictionaryShared::check_for_exclusion(InstanceKlass* k, DumpTimeClassInfo* info) {
+  if (MetaspaceShared::is_in_shared_metaspace(k)) {
+    // We have reached a super type that's already in the base archive. Treat it
+    // as "not excluded".
+    assert(DynamicDumpSharedSpaces, "must be");
+    return false;
+  }
+
+  if (info == NULL) {
+    info = _dumptime_table->get(k);
+    assert(info != NULL, "supertypes of any classes in _dumptime_table must either be shared, or must also be in _dumptime_table");
+  }
+
+  if (!info->has_checked_exclusion()) {
+    if (check_for_exclusion_impl(k)) {
+      info->set_excluded();
+    }
+    info->set_has_checked_exclusion();
+  }
+
+  return info->is_excluded();
+}
+
+// Returns true so the caller can do:    return warn_excluded(".....");
+bool SystemDictionaryShared::warn_excluded(InstanceKlass* k, const char* reason) {
+  ResourceMark rm;
+  log_warning(cds)("Skipping %s: %s", k->name()->as_C_string(), reason);
+  return true;
+}
+
+bool SystemDictionaryShared::is_jfr_event_class(InstanceKlass *k) {
+  while (k) {
+    if (k->name()->equals("jdk/internal/event/Event")) {
+      return true;
+    }
+    k = k->java_super();
+  }
+  return false;
+}
+
+bool SystemDictionaryShared::is_registered_lambda_proxy_class(InstanceKlass* ik) {
+  DumpTimeClassInfo* info = _dumptime_table->get(ik);
+  return (info != NULL) ? info->_is_archived_lambda_proxy : false;
+}
+
+bool SystemDictionaryShared::is_early_klass(InstanceKlass* ik) {
+  DumpTimeClassInfo* info = _dumptime_table->get(ik);
+  return (info != NULL) ? info->is_early_klass() : false;
+}
+
+bool SystemDictionaryShared::is_hidden_lambda_proxy(InstanceKlass* ik) {
+  assert(ik->is_shared(), "applicable to only a shared class");
+  if (ik->is_hidden()) {
+    return true;
   } else {
-    lambda_info->add_proxy_klass(proxy_klass);
+    return false;
   }
 }
 
-class RunTimeSharedClassInfo {
-public:
-  struct CrcInfo {
-    int _clsfile_size;
-    int _clsfile_crc32;
-  };
-
-  // This is different than  DumpTimeSharedClassInfo::DTVerifierConstraint. We use
-  // u4 instead of Symbol* to save space on 64-bit CPU.
-  struct RTVerifierConstraint {
-    u4 _name;
-    u4 _from_name;
-    Symbol* name() { return (Symbol*)(SharedBaseAddress + _name);}
-    Symbol* from_name() { return (Symbol*)(SharedBaseAddress + _from_name); }
-  };
-
-  struct RTLoaderConstraint {
-    u4   _name;
-    char _loader_type1;
-    char _loader_type2;
-    Symbol* constraint_name() {
-      return (Symbol*)(SharedBaseAddress + _name);
-    }
-  };
-
-  InstanceKlass* _klass;
-  int _num_verifier_constraints;
-  int _num_loader_constraints;
-
-  // optional CrcInfo              _crc;  (only for UNREGISTERED classes)
-  // optional InstanceKlass*       _nest_host
-  // optional RTLoaderConstraint   _loader_constraint_types[_num_loader_constraints]
-  // optional RTVerifierConstraint _verifier_constraints[_num_verifier_constraints]
-  // optional char                 _verifier_constraint_flags[_num_verifier_constraints]
-
-private:
-  static size_t header_size_size() {
-    return sizeof(RunTimeSharedClassInfo);
+bool SystemDictionaryShared::check_for_exclusion_impl(InstanceKlass* k) {
+  if (k->is_in_error_state()) {
+    return warn_excluded(k, "In error state");
   }
-  static size_t crc_size(InstanceKlass* klass) {
-    if (!SystemDictionaryShared::is_builtin(klass)) {
-      return sizeof(CrcInfo);
+  if (has_been_redefined(k)) {
+    return warn_excluded(k, "Has been redefined");
+  }
+  if (!k->is_hidden() && k->shared_classpath_index() < 0 && is_builtin(k)) {
+    // These are classes loaded from unsupported locations (such as those loaded by JVMTI native
+    // agent during dump time).
+    return warn_excluded(k, "Unsupported location");
+  }
+  if (k->signers() != NULL) {
+    // We cannot include signed classes in the archive because the certificates
+    // used during dump time may be different than those used during
+    // runtime (due to expiration, etc).
+    return warn_excluded(k, "Signed JAR");
+  }
+  if (is_jfr_event_class(k)) {
+    // We cannot include JFR event classes because they need runtime-specific
+    // instrumentation in order to work with -XX:FlightRecorderOptions:retransform=false.
+    // There are only a small number of these classes, so it's not worthwhile to
+    // support them and make CDS more complicated.
+    return warn_excluded(k, "JFR event class");
+  }
+  if (k->init_state() < InstanceKlass::linked) {
+    // In CDS dumping, we will attempt to link all classes. Those that fail to link will
+    // be recorded in DumpTimeClassInfo.
+    Arguments::assert_is_dumping_archive();
+
+    // TODO -- rethink how this can be handled.
+    // We should try to link ik, however, we can't do it here because
+    // 1. We are at VM exit
+    // 2. linking a class may cause other classes to be loaded, which means
+    //    a custom ClassLoader.loadClass() may be called, at a point where the
+    //    class loader doesn't expect it.
+    if (has_class_failed_verification(k)) {
+      return warn_excluded(k, "Failed verification");
     } else {
-      return 0;
-    }
-  }
-  static size_t verifier_constraints_size(int num_verifier_constraints) {
-    return sizeof(RTVerifierConstraint) * num_verifier_constraints;
-  }
-  static size_t verifier_constraint_flags_size(int num_verifier_constraints) {
-    return sizeof(char) * num_verifier_constraints;
-  }
-  static size_t loader_constraints_size(int num_loader_constraints) {
-    return sizeof(RTLoaderConstraint) * num_loader_constraints;
-  }
-  static size_t nest_host_size(InstanceKlass* klass) {
-    if (klass->is_hidden()) {
-      return sizeof(InstanceKlass*);
-    } else {
-      return 0;
-    }
-  }
-
-public:
-  static size_t byte_size(InstanceKlass* klass, int num_verifier_constraints, int num_loader_constraints) {
-    return header_size_size() +
-           crc_size(klass) +
-           nest_host_size(klass) +
-           loader_constraints_size(num_loader_constraints) +
-           verifier_constraints_size(num_verifier_constraints) +
-           verifier_constraint_flags_size(num_verifier_constraints);
-  }
-
-private:
-  size_t crc_offset() const {
-    return header_size_size();
-  }
-
-  size_t nest_host_offset() const {
-      return crc_offset() + crc_size(_klass);
-  }
-
-  size_t loader_constraints_offset() const  {
-    return nest_host_offset() + nest_host_size(_klass);
-  }
-  size_t verifier_constraints_offset() const {
-    return loader_constraints_offset() + loader_constraints_size(_num_loader_constraints);
-  }
-  size_t verifier_constraint_flags_offset() const {
-    return verifier_constraints_offset() + verifier_constraints_size(_num_verifier_constraints);
-  }
-
-  void check_verifier_constraint_offset(int i) const {
-    assert(0 <= i && i < _num_verifier_constraints, "sanity");
-  }
-
-  void check_loader_constraint_offset(int i) const {
-    assert(0 <= i && i < _num_loader_constraints, "sanity");
-  }
-
-public:
-  CrcInfo* crc() const {
-    assert(crc_size(_klass) > 0, "must be");
-    return (CrcInfo*)(address(this) + crc_offset());
-  }
-  RTVerifierConstraint* verifier_constraints() {
-    assert(_num_verifier_constraints > 0, "sanity");
-    return (RTVerifierConstraint*)(address(this) + verifier_constraints_offset());
-  }
-  RTVerifierConstraint* verifier_constraint_at(int i) {
-    check_verifier_constraint_offset(i);
-    return verifier_constraints() + i;
-  }
-
-  char* verifier_constraint_flags() {
-    assert(_num_verifier_constraints > 0, "sanity");
-    return (char*)(address(this) + verifier_constraint_flags_offset());
-  }
-
-  InstanceKlass** nest_host_addr() {
-    assert(_klass->is_hidden(), "sanity");
-    return (InstanceKlass**)(address(this) + nest_host_offset());
-  }
-  InstanceKlass* nest_host() {
-    return *nest_host_addr();
-  }
-  void set_nest_host(InstanceKlass* k) {
-    *nest_host_addr() = k;
-    ArchivePtrMarker::mark_pointer((address*)nest_host_addr());
-  }
-
-  RTLoaderConstraint* loader_constraints() {
-    assert(_num_loader_constraints > 0, "sanity");
-    return (RTLoaderConstraint*)(address(this) + loader_constraints_offset());
-  }
-
-  RTLoaderConstraint* loader_constraint_at(int i) {
-    check_loader_constraint_offset(i);
-    return loader_constraints() + i;
-  }
-
-  void init(DumpTimeSharedClassInfo& info) {
-    ArchiveBuilder* builder = ArchiveBuilder::current();
-    assert(builder->is_in_buffer_space(info._klass), "must be");
-    _klass = info._klass;
-    if (!SystemDictionaryShared::is_builtin(_klass)) {
-      CrcInfo* c = crc();
-      c->_clsfile_size = info._clsfile_size;
-      c->_clsfile_crc32 = info._clsfile_crc32;
-    }
-    _num_verifier_constraints = info.num_verifier_constraints();
-    _num_loader_constraints   = info.num_loader_constraints();
-    int i;
-    if (_num_verifier_constraints > 0) {
-      RTVerifierConstraint* vf_constraints = verifier_constraints();
-      char* flags = verifier_constraint_flags();
-      for (i = 0; i < _num_verifier_constraints; i++) {
-        vf_constraints[i]._name      = builder->any_to_offset_u4(info._verifier_constraints->at(i)._name);
-        vf_constraints[i]._from_name = builder->any_to_offset_u4(info._verifier_constraints->at(i)._from_name);
-      }
-      for (i = 0; i < _num_verifier_constraints; i++) {
-        flags[i] = info._verifier_constraint_flags->at(i);
+      if (k->can_be_verified_at_dumptime()) {
+        return warn_excluded(k, "Not linked");
       }
     }
-
-    if (_num_loader_constraints > 0) {
-      RTLoaderConstraint* ld_constraints = loader_constraints();
-      for (i = 0; i < _num_loader_constraints; i++) {
-        ld_constraints[i]._name = builder->any_to_offset_u4(info._loader_constraints->at(i)._name);
-        ld_constraints[i]._loader_type1 = info._loader_constraints->at(i)._loader_type1;
-        ld_constraints[i]._loader_type2 = info._loader_constraints->at(i)._loader_type2;
-      }
-    }
-
-    if (_klass->is_hidden()) {
-      InstanceKlass* n_h = info.nest_host();
-      set_nest_host(n_h);
-    }
-    ArchivePtrMarker::mark_pointer(&_klass);
+  }
+  if (DynamicDumpSharedSpaces && k->major_version() < 50 /*JAVA_6_VERSION*/) {
+    // In order to support old classes during dynamic dump, class rewriting needs to
+    // be reverted. This would result in more complex code and testing but not much gain.
+    ResourceMark rm;
+    log_warning(cds)("Pre JDK 6 class not supported by CDS: %u.%u %s",
+                     k->major_version(),  k->minor_version(), k->name()->as_C_string());
+    return true;
   }
 
-  bool matches(int clsfile_size, int clsfile_crc32) const {
-    return crc()->_clsfile_size  == clsfile_size &&
-           crc()->_clsfile_crc32 == clsfile_crc32;
+  if (!k->can_be_verified_at_dumptime() && k->is_linked()) {
+    return warn_excluded(k, "Old class has been linked");
   }
 
-  char verifier_constraint_flag(int i) {
-    check_verifier_constraint_offset(i);
-    return verifier_constraint_flags()[i];
+  if (k->is_hidden() && !is_registered_lambda_proxy_class(k)) {
+    ResourceMark rm;
+    log_debug(cds)("Skipping %s: Hidden class", k->name()->as_C_string());
+    return true;
   }
 
-private:
-  // ArchiveBuilder::make_shallow_copy() has reserved a pointer immediately
-  // before archived InstanceKlasses. We can use this slot to do a quick
-  // lookup of InstanceKlass* -> RunTimeSharedClassInfo* without
-  // building a new hashtable.
-  //
-  //  info_pointer_addr(klass) --> 0x0100   RunTimeSharedClassInfo*
-  //  InstanceKlass* klass     --> 0x0108   <C++ vtbl>
-  //                               0x0110   fields from Klass ...
-  static RunTimeSharedClassInfo** info_pointer_addr(InstanceKlass* klass) {
-    return &((RunTimeSharedClassInfo**)klass)[-1];
+  InstanceKlass* super = k->java_super();
+  if (super != NULL && check_for_exclusion(super, NULL)) {
+    ResourceMark rm;
+    log_warning(cds)("Skipping %s: super class %s is excluded", k->name()->as_C_string(), super->name()->as_C_string());
+    return true;
   }
 
-public:
-  static RunTimeSharedClassInfo* get_for(InstanceKlass* klass) {
-    assert(klass->is_shared(), "don't call for non-shared class");
-    return *info_pointer_addr(klass);
-  }
-  static void set_for(InstanceKlass* klass, RunTimeSharedClassInfo* record) {
-    assert(ArchiveBuilder::current()->is_in_buffer_space(klass), "must be");
-    assert(ArchiveBuilder::current()->is_in_buffer_space(record), "must be");
-    *info_pointer_addr(klass) = record;
-    ArchivePtrMarker::mark_pointer(info_pointer_addr(klass));
-  }
-
-  // Used by RunTimeSharedDictionary to implement OffsetCompactHashtable::EQUALS
-  static inline bool EQUALS(
-       const RunTimeSharedClassInfo* value, Symbol* key, int len_unused) {
-    return (value->_klass->name() == key);
-  }
-};
-
-class RunTimeSharedDictionary : public OffsetCompactHashtable<
-  Symbol*,
-  const RunTimeSharedClassInfo*,
-  RunTimeSharedClassInfo::EQUALS> {};
-
-static DumpTimeSharedClassTable* _dumptime_table = NULL;
-// SystemDictionaries in the base layer static archive
-static RunTimeSharedDictionary _builtin_dictionary;
-static RunTimeSharedDictionary _unregistered_dictionary;
-// SystemDictionaries in the top layer dynamic archive
-static RunTimeSharedDictionary _dynamic_builtin_dictionary;
-static RunTimeSharedDictionary _dynamic_unregistered_dictionary;
-
-void SystemDictionaryShared::atomic_set_array_index(OopHandle array, int index, oop o) {
-  // Benign race condition:  array.obj_at(index) may already be filled in.
-  // The important thing here is that all threads pick up the same result.
-  // It doesn't matter which racing thread wins, as long as only one
-  // result is used by all threads, and all future queries.
-  ((objArrayOop)array.resolve())->atomic_compare_exchange_oop(index, o, NULL);
-}
-
-Handle SystemDictionaryShared::create_jar_manifest(const char* manifest_chars, size_t size, TRAPS) {
-  typeArrayOop buf = oopFactory::new_byteArray((int)size, CHECK_NH);
-  typeArrayHandle bufhandle(THREAD, buf);
-  ArrayAccess<>::arraycopy_from_native(reinterpret_cast<const jbyte*>(manifest_chars),
-                                         buf, typeArrayOopDesc::element_offset<jbyte>(0), size);
-  Handle bais = JavaCalls::construct_new_instance(vmClasses::ByteArrayInputStream_klass(),
-                      vmSymbols::byte_array_void_signature(),
-                      bufhandle, CHECK_NH);
-  // manifest = new Manifest(ByteArrayInputStream)
-  Handle manifest = JavaCalls::construct_new_instance(vmClasses::Jar_Manifest_klass(),
-                      vmSymbols::input_stream_void_signature(),
-                      bais, CHECK_NH);
-  return manifest;
-}
-
-oop SystemDictionaryShared::shared_protection_domain(int index) {
-  return ((objArrayOop)_shared_protection_domains.resolve())->obj_at(index);
-}
-
-oop SystemDictionaryShared::shared_jar_url(int index) {
-  return ((objArrayOop)_shared_jar_urls.resolve())->obj_at(index);
-}
-
-oop SystemDictionaryShared::shared_jar_manifest(int index) {
-  return ((objArrayOop)_shared_jar_manifests.resolve())->obj_at(index);
-}
-
-Handle SystemDictionaryShared::get_shared_jar_manifest(int shared_path_index, TRAPS) {
-  Handle manifest ;
-  if (shared_jar_manifest(shared_path_index) == NULL) {
-    SharedClassPathEntry* ent = FileMapInfo::shared_path(shared_path_index);
-    size_t size = (size_t)ent->manifest_size();
-    if (size == 0) {
-      return Handle();
-    }
-
-    // ByteArrayInputStream bais = new ByteArrayInputStream(buf);
-    const char* src = ent->manifest();
-    assert(src != NULL, "No Manifest data");
-    manifest = create_jar_manifest(src, size, CHECK_NH);
-    atomic_set_shared_jar_manifest(shared_path_index, manifest());
-  }
-  manifest = Handle(THREAD, shared_jar_manifest(shared_path_index));
-  assert(manifest.not_null(), "sanity");
-  return manifest;
-}
-
-Handle SystemDictionaryShared::get_shared_jar_url(int shared_path_index, TRAPS) {
-  Handle url_h;
-  if (shared_jar_url(shared_path_index) == NULL) {
-    JavaValue result(T_OBJECT);
-    const char* path = FileMapInfo::shared_path_name(shared_path_index);
-    Handle path_string = java_lang_String::create_from_str(path, CHECK_(url_h));
-    Klass* classLoaders_klass =
-        vmClasses::jdk_internal_loader_ClassLoaders_klass();
-    JavaCalls::call_static(&result, classLoaders_klass,
-                           vmSymbols::toFileURL_name(),
-                           vmSymbols::toFileURL_signature(),
-                           path_string, CHECK_(url_h));
-
-    atomic_set_shared_jar_url(shared_path_index, result.get_oop());
-  }
-
-  url_h = Handle(THREAD, shared_jar_url(shared_path_index));
-  assert(url_h.not_null(), "sanity");
-  return url_h;
-}
-
-Handle SystemDictionaryShared::get_package_name(Symbol* class_name, TRAPS) {
-  ResourceMark rm(THREAD);
-  Handle pkgname_string;
-  TempNewSymbol pkg = ClassLoader::package_from_class_name(class_name);
-  if (pkg != NULL) { // Package prefix found
-    const char* pkgname = pkg->as_klass_external_name();
-    pkgname_string = java_lang_String::create_from_str(pkgname,
-                                                       CHECK_(pkgname_string));
-  }
-  return pkgname_string;
-}
-
-// Define Package for shared app classes from JAR file and also checks for
-// package sealing (all done in Java code)
-// See http://docs.oracle.com/javase/tutorial/deployment/jar/sealman.html
-void SystemDictionaryShared::define_shared_package(Symbol*  class_name,
-                                                   Handle class_loader,
-                                                   Handle manifest,
-                                                   Handle url,
-                                                   TRAPS) {
-  assert(SystemDictionary::is_system_class_loader(class_loader()), "unexpected class loader");
-  // get_package_name() returns a NULL handle if the class is in unnamed package
-  Handle pkgname_string = get_package_name(class_name, CHECK);
-  if (pkgname_string.not_null()) {
-    Klass* app_classLoader_klass = vmClasses::jdk_internal_loader_ClassLoaders_AppClassLoader_klass();
-    JavaValue result(T_OBJECT);
-    JavaCallArguments args(3);
-    args.set_receiver(class_loader);
-    args.push_oop(pkgname_string);
-    args.push_oop(manifest);
-    args.push_oop(url);
-    JavaCalls::call_virtual(&result, app_classLoader_klass,
-                            vmSymbols::defineOrCheckPackage_name(),
-                            vmSymbols::defineOrCheckPackage_signature(),
-                            &args,
-                            CHECK);
-  }
-}
-
-// Get the ProtectionDomain associated with the CodeSource from the classloader.
-Handle SystemDictionaryShared::get_protection_domain_from_classloader(Handle class_loader,
-                                                                      Handle url, TRAPS) {
-  // CodeSource cs = new CodeSource(url, null);
-  Handle cs = JavaCalls::construct_new_instance(vmClasses::CodeSource_klass(),
-                  vmSymbols::url_code_signer_array_void_signature(),
-                  url, Handle(), CHECK_NH);
-
-  // protection_domain = SecureClassLoader.getProtectionDomain(cs);
-  Klass* secureClassLoader_klass = vmClasses::SecureClassLoader_klass();
-  JavaValue obj_result(T_OBJECT);
-  JavaCalls::call_virtual(&obj_result, class_loader, secureClassLoader_klass,
-                          vmSymbols::getProtectionDomain_name(),
-                          vmSymbols::getProtectionDomain_signature(),
-                          cs, CHECK_NH);
-  return Handle(THREAD, obj_result.get_oop());
-}
-
-// Returns the ProtectionDomain associated with the JAR file identified by the url.
-Handle SystemDictionaryShared::get_shared_protection_domain(Handle class_loader,
-                                                            int shared_path_index,
-                                                            Handle url,
-                                                            TRAPS) {
-  Handle protection_domain;
-  if (shared_protection_domain(shared_path_index) == NULL) {
-    Handle pd = get_protection_domain_from_classloader(class_loader, url, THREAD);
-    atomic_set_shared_protection_domain(shared_path_index, pd());
-  }
-
-  // Acquire from the cache because if another thread beats the current one to
-  // set the shared protection_domain and the atomic_set fails, the current thread
-  // needs to get the updated protection_domain from the cache.
-  protection_domain = Handle(THREAD, shared_protection_domain(shared_path_index));
-  assert(protection_domain.not_null(), "sanity");
-  return protection_domain;
-}
-
-// Returns the ProtectionDomain associated with the moduleEntry.
-Handle SystemDictionaryShared::get_shared_protection_domain(Handle class_loader,
-                                                            ModuleEntry* mod, TRAPS) {
-  ClassLoaderData *loader_data = mod->loader_data();
-  if (mod->shared_protection_domain() == NULL) {
-    Symbol* location = mod->location();
-    if (location != NULL) {
-      Handle location_string = java_lang_String::create_from_symbol(
-                                     location, CHECK_NH);
-      Handle url;
-      JavaValue result(T_OBJECT);
-      if (location->starts_with("jrt:/")) {
-        url = JavaCalls::construct_new_instance(vmClasses::URL_klass(),
-                                                vmSymbols::string_void_signature(),
-                                                location_string, CHECK_NH);
-      } else {
-        Klass* classLoaders_klass =
-          vmClasses::jdk_internal_loader_ClassLoaders_klass();
-        JavaCalls::call_static(&result, classLoaders_klass, vmSymbols::toFileURL_name(),
-                               vmSymbols::toFileURL_signature(),
-                               location_string, CHECK_NH);
-        url = Handle(THREAD, result.get_oop());
-      }
-
-      Handle pd = get_protection_domain_from_classloader(class_loader, url,
-                                                         CHECK_NH);
-      mod->set_shared_protection_domain(loader_data, pd);
+  Array<InstanceKlass*>* interfaces = k->local_interfaces();
+  int len = interfaces->length();
+  for (int i = 0; i < len; i++) {
+    InstanceKlass* intf = interfaces->at(i);
+    if (check_for_exclusion(intf, NULL)) {
+      log_warning(cds)("Skipping %s: interface %s is excluded", k->name()->as_C_string(), intf->name()->as_C_string());
+      return true;
     }
   }
 
-  Handle protection_domain(THREAD, mod->shared_protection_domain());
-  assert(protection_domain.not_null(), "sanity");
-  return protection_domain;
-}
-
-// Initializes the java.lang.Package and java.security.ProtectionDomain objects associated with
-// the given InstanceKlass.
-// Returns the ProtectionDomain for the InstanceKlass.
-Handle SystemDictionaryShared::init_security_info(Handle class_loader, InstanceKlass* ik, PackageEntry* pkg_entry, TRAPS) {
-  Handle pd;
-
-  if (ik != NULL) {
-    int index = ik->shared_classpath_index();
-    assert(index >= 0, "Sanity");
-    SharedClassPathEntry* ent = FileMapInfo::shared_path(index);
-    Symbol* class_name = ik->name();
-
-    if (ent->is_modules_image()) {
-      // For shared app/platform classes originated from the run-time image:
-      //   The ProtectionDomains are cached in the corresponding ModuleEntries
-      //   for fast access by the VM.
-      // all packages from module image are already created during VM bootstrap in
-      // Modules::define_module().
-      assert(pkg_entry != NULL, "archived class in module image cannot be from unnamed package");
-      ModuleEntry* mod_entry = pkg_entry->module();
-      pd = get_shared_protection_domain(class_loader, mod_entry, CHECK_(pd));
-    } else {
-      // For shared app/platform classes originated from JAR files on the class path:
-      //   Each of the 3 SystemDictionaryShared::_shared_xxx arrays has the same length
-      //   as the shared classpath table in the shared archive (see
-      //   FileMap::_shared_path_table in filemap.hpp for details).
-      //
-      //   If a shared InstanceKlass k is loaded from the class path, let
-      //
-      //     index = k->shared_classpath_index():
-      //
-      //   FileMap::_shared_path_table[index] identifies the JAR file that contains k.
-      //
-      //   k's protection domain is:
-      //
-      //     ProtectionDomain pd = _shared_protection_domains[index];
-      //
-      //   and k's Package is initialized using
-      //
-      //     manifest = _shared_jar_manifests[index];
-      //     url = _shared_jar_urls[index];
-      //     define_shared_package(class_name, class_loader, manifest, url, CHECK_(pd));
-      //
-      //   Note that if an element of these 3 _shared_xxx arrays is NULL, it will be initialized by
-      //   the corresponding SystemDictionaryShared::get_shared_xxx() function.
-      Handle manifest = get_shared_jar_manifest(index, CHECK_(pd));
-      Handle url = get_shared_jar_url(index, CHECK_(pd));
-      int index_offset = index - ClassLoaderExt::app_class_paths_start_index();
-      if (index_offset < PackageEntry::max_index_for_defined_in_class_path()) {
-        if (pkg_entry == NULL || !pkg_entry->is_defined_by_cds_in_class_path(index_offset)) {
-          // define_shared_package only needs to be called once for each package in a jar specified
-          // in the shared class path.
-          define_shared_package(class_name, class_loader, manifest, url, CHECK_(pd));
-          if (pkg_entry != NULL) {
-            pkg_entry->set_defined_by_cds_in_class_path(index_offset);
-          }
-        }
-      } else {
-        define_shared_package(class_name, class_loader, manifest, url, CHECK_(pd));
-      }
-      pd = get_shared_protection_domain(class_loader, index, url, CHECK_(pd));
-    }
-  }
-  return pd;
+  return false; // false == k should NOT be excluded
 }
 
 bool SystemDictionaryShared::is_sharing_possible(ClassLoaderData* loader_data) {
@@ -1026,145 +427,6 @@ InstanceKlass* SystemDictionaryShared::find_or_load_shared_class(
   return k;
 }
 
-PackageEntry* SystemDictionaryShared::get_package_entry_from_class(InstanceKlass* ik, Handle class_loader) {
-  PackageEntry* pkg_entry = ik->package();
-  if (MetaspaceShared::use_full_module_graph() && ik->is_shared() && pkg_entry != NULL) {
-    assert(MetaspaceShared::is_in_shared_metaspace(pkg_entry), "must be");
-    assert(!ik->is_shared_unregistered_class(), "unexpected archived package entry for an unregistered class");
-    assert(ik->module()->is_named(), "unexpected archived package entry for a class in an unnamed module");
-    return pkg_entry;
-  }
-  TempNewSymbol pkg_name = ClassLoader::package_from_class_name(ik->name());
-  if (pkg_name != NULL) {
-    pkg_entry = class_loader_data(class_loader)->packages()->lookup_only(pkg_name);
-  } else {
-    pkg_entry = NULL;
-  }
-  return pkg_entry;
-}
-
-InstanceKlass* SystemDictionaryShared::load_shared_class_for_builtin_loader(
-                 Symbol* class_name, Handle class_loader, TRAPS) {
-  assert(UseSharedSpaces, "must be");
-  InstanceKlass* ik = find_builtin_class(class_name);
-
-  if (ik != NULL && !ik->shared_loading_failed()) {
-    if ((SystemDictionary::is_system_class_loader(class_loader()) && ik->is_shared_app_class())  ||
-        (SystemDictionary::is_platform_class_loader(class_loader()) && ik->is_shared_platform_class())) {
-      SharedClassLoadingMark slm(THREAD, ik);
-      PackageEntry* pkg_entry = get_package_entry_from_class(ik, class_loader);
-      Handle protection_domain =
-        SystemDictionaryShared::init_security_info(class_loader, ik, pkg_entry, CHECK_NULL);
-      return load_shared_class(ik, class_loader, protection_domain, NULL, pkg_entry, THREAD);
-    }
-  }
-  return NULL;
-}
-
-void SystemDictionaryShared::allocate_shared_protection_domain_array(int size, TRAPS) {
-  if (_shared_protection_domains.resolve() == NULL) {
-    oop spd = oopFactory::new_objArray(
-        vmClasses::ProtectionDomain_klass(), size, CHECK);
-    _shared_protection_domains = OopHandle(Universe::vm_global(), spd);
-  }
-}
-
-void SystemDictionaryShared::allocate_shared_jar_url_array(int size, TRAPS) {
-  if (_shared_jar_urls.resolve() == NULL) {
-    oop sju = oopFactory::new_objArray(
-        vmClasses::URL_klass(), size, CHECK);
-    _shared_jar_urls = OopHandle(Universe::vm_global(), sju);
-  }
-}
-
-void SystemDictionaryShared::allocate_shared_jar_manifest_array(int size, TRAPS) {
-  if (_shared_jar_manifests.resolve() == NULL) {
-    oop sjm = oopFactory::new_objArray(
-        vmClasses::Jar_Manifest_klass(), size, CHECK);
-    _shared_jar_manifests = OopHandle(Universe::vm_global(), sjm);
-  }
-}
-
-void SystemDictionaryShared::allocate_shared_data_arrays(int size, TRAPS) {
-  allocate_shared_protection_domain_array(size, CHECK);
-  allocate_shared_jar_url_array(size, CHECK);
-  allocate_shared_jar_manifest_array(size, CHECK);
-}
-
-// This function is called for loading only UNREGISTERED classes
-InstanceKlass* SystemDictionaryShared::lookup_from_stream(Symbol* class_name,
-                                                          Handle class_loader,
-                                                          Handle protection_domain,
-                                                          const ClassFileStream* cfs,
-                                                          TRAPS) {
-  if (!UseSharedSpaces) {
-    return NULL;
-  }
-  if (class_name == NULL) {  // don't do this for hidden classes
-    return NULL;
-  }
-  if (class_loader.is_null() ||
-      SystemDictionary::is_system_class_loader(class_loader()) ||
-      SystemDictionary::is_platform_class_loader(class_loader())) {
-    // Do nothing for the BUILTIN loaders.
-    return NULL;
-  }
-
-  const RunTimeSharedClassInfo* record = find_record(&_unregistered_dictionary, &_dynamic_unregistered_dictionary, class_name);
-  if (record == NULL) {
-    return NULL;
-  }
-
-  int clsfile_size  = cfs->length();
-  int clsfile_crc32 = ClassLoader::crc32(0, (const char*)cfs->buffer(), cfs->length());
-
-  if (!record->matches(clsfile_size, clsfile_crc32)) {
-    return NULL;
-  }
-
-  return acquire_class_for_current_thread(record->_klass, class_loader,
-                                          protection_domain, cfs,
-                                          THREAD);
-}
-
-InstanceKlass* SystemDictionaryShared::acquire_class_for_current_thread(
-                   InstanceKlass *ik,
-                   Handle class_loader,
-                   Handle protection_domain,
-                   const ClassFileStream *cfs,
-                   TRAPS) {
-  ClassLoaderData* loader_data = ClassLoaderData::class_loader_data(class_loader());
-
-  {
-    MutexLocker mu(THREAD, SharedDictionary_lock);
-    if (ik->class_loader_data() != NULL) {
-      //    ik is already loaded (by this loader or by a different loader)
-      // or ik is being loaded by a different thread (by this loader or by a different loader)
-      return NULL;
-    }
-
-    // No other thread has acquired this yet, so give it to *this thread*
-    ik->set_class_loader_data(loader_data);
-  }
-
-  // No longer holding SharedDictionary_lock
-  // No need to lock, as <ik> can be held only by a single thread.
-  loader_data->add_class(ik);
-
-  // Get the package entry.
-  PackageEntry* pkg_entry = get_package_entry_from_class(ik, class_loader);
-
-  // Load and check super/interfaces, restore unsharable info
-  InstanceKlass* shared_klass = load_shared_class(ik, class_loader, protection_domain,
-                                                  cfs, pkg_entry, THREAD);
-  if (shared_klass == NULL || HAS_PENDING_EXCEPTION) {
-    // TODO: clean up <ik> so it can be used again
-    return NULL;
-  }
-
-  return shared_klass;
-}
-
 class UnregisteredClassesTable : public ResourceHashtable<
   Symbol*, InstanceKlass*,
   primitive_hash<Symbol*>,
@@ -1243,28 +505,10 @@ InstanceKlass* SystemDictionaryShared::lookup_super_for_unregistered_class(
   }
 }
 
-void SystemDictionaryShared::start_dumping() {
-  MutexLocker ml(DumpTimeTable_lock, Mutex::_no_safepoint_check_flag);
-  _dump_in_progress = true;
-}
-
-DumpTimeSharedClassInfo* SystemDictionaryShared::find_or_allocate_info_for(InstanceKlass* k) {
-  MutexLocker ml(DumpTimeTable_lock, Mutex::_no_safepoint_check_flag);
-  return find_or_allocate_info_for_locked(k);
-}
-
-DumpTimeSharedClassInfo* SystemDictionaryShared::find_or_allocate_info_for_locked(InstanceKlass* k) {
-  assert_lock_strong(DumpTimeTable_lock);
-  if (_dumptime_table == NULL) {
-    _dumptime_table = new (ResourceObj::C_HEAP, mtClass)DumpTimeSharedClassTable();
-  }
-  return _dumptime_table->find_or_allocate_info_for(k, _dump_in_progress);
-}
-
 void SystemDictionaryShared::set_shared_class_misc_info(InstanceKlass* k, ClassFileStream* cfs) {
   Arguments::assert_is_dumping_archive();
   assert(!is_builtin(k), "must be unregistered class");
-  DumpTimeSharedClassInfo* info = find_or_allocate_info_for(k);
+  DumpTimeClassInfo* info = find_or_allocate_info_for(k);
   if (info != NULL) {
     info->_clsfile_size  = cfs->length();
     info->_clsfile_crc32 = ClassLoader::crc32(0, (const char*)cfs->buffer(), cfs->length());
@@ -1277,13 +521,13 @@ void SystemDictionaryShared::init_dumptime_info(InstanceKlass* k) {
 
 void SystemDictionaryShared::remove_dumptime_info(InstanceKlass* k) {
   MutexLocker ml(DumpTimeTable_lock, Mutex::_no_safepoint_check_flag);
-  DumpTimeSharedClassInfo* p = _dumptime_table->get(k);
+  DumpTimeClassInfo* p = _dumptime_table->get(k);
   if (p == NULL) {
     return;
   }
   if (p->_verifier_constraints != NULL) {
     for (int i = 0; i < p->_verifier_constraints->length(); i++) {
-      DumpTimeSharedClassInfo::DTVerifierConstraint constraint = p->_verifier_constraints->at(i);
+      DumpTimeClassInfo::DTVerifierConstraint constraint = p->_verifier_constraints->at(i);
       if (constraint._name != NULL ) {
         constraint._name->decrement_refcount();
       }
@@ -1291,19 +535,19 @@ void SystemDictionaryShared::remove_dumptime_info(InstanceKlass* k) {
         constraint._from_name->decrement_refcount();
       }
     }
-    FREE_C_HEAP_ARRAY(DumpTimeSharedClassInfo::DTVerifierConstraint, p->_verifier_constraints);
+    FREE_C_HEAP_ARRAY(DumpTimeClassInfo::DTVerifierConstraint, p->_verifier_constraints);
     p->_verifier_constraints = NULL;
     FREE_C_HEAP_ARRAY(char, p->_verifier_constraint_flags);
     p->_verifier_constraint_flags = NULL;
   }
   if (p->_loader_constraints != NULL) {
     for (int i = 0; i < p->_loader_constraints->length(); i++) {
-      DumpTimeSharedClassInfo::DTLoaderConstraint ld =  p->_loader_constraints->at(i);
+      DumpTimeClassInfo::DTLoaderConstraint ld =  p->_loader_constraints->at(i);
       if (ld._name != NULL) {
         ld._name->decrement_refcount();
       }
     }
-    FREE_C_HEAP_ARRAY(DumpTimeSharedClassInfo::DTLoaderConstraint, p->_loader_constraints);
+    FREE_C_HEAP_ARRAY(DumpTimeClassInfo::DTLoaderConstraint, p->_loader_constraints);
     p->_loader_constraints = NULL;
   }
   _dumptime_table->remove(k);
@@ -1324,65 +568,6 @@ void SystemDictionaryShared::handle_class_unloading(InstanceKlass* klass) {
   }
 }
 
-bool SystemDictionaryShared::is_jfr_event_class(InstanceKlass *k) {
-  while (k) {
-    if (k->name()->equals("jdk/internal/event/Event")) {
-      return true;
-    }
-    k = k->java_super();
-  }
-  return false;
-}
-
-bool SystemDictionaryShared::is_registered_lambda_proxy_class(InstanceKlass* ik) {
-  DumpTimeSharedClassInfo* info = _dumptime_table->get(ik);
-  return (info != NULL) ? info->_is_archived_lambda_proxy : false;
-}
-
-bool SystemDictionaryShared::is_hidden_lambda_proxy(InstanceKlass* ik) {
-  assert(ik->is_shared(), "applicable to only a shared class");
-  if (ik->is_hidden()) {
-    return true;
-  } else {
-    return false;
-  }
-}
-
-bool SystemDictionaryShared::is_early_klass(InstanceKlass* ik) {
-  DumpTimeSharedClassInfo* info = _dumptime_table->get(ik);
-  return (info != NULL) ? info->is_early_klass() : false;
-}
-
-// Returns true so the caller can do:    return warn_excluded(".....");
-bool SystemDictionaryShared::warn_excluded(InstanceKlass* k, const char* reason) {
-  ResourceMark rm;
-  log_warning(cds)("Skipping %s: %s", k->name()->as_C_string(), reason);
-  return true;
-}
-
-bool SystemDictionaryShared::check_for_exclusion(InstanceKlass* k, DumpTimeSharedClassInfo* info) {
-  if (MetaspaceShared::is_in_shared_metaspace(k)) {
-    // We have reached a super type that's already in the base archive. Treat it
-    // as "not excluded".
-    assert(DynamicDumpSharedSpaces, "must be");
-    return false;
-  }
-
-  if (info == NULL) {
-    info = _dumptime_table->get(k);
-    assert(info != NULL, "supertypes of any classes in _dumptime_table must either be shared, or must also be in _dumptime_table");
-  }
-
-  if (!info->has_checked_exclusion()) {
-    if (check_for_exclusion_impl(k)) {
-      info->set_excluded();
-    }
-    info->set_has_checked_exclusion();
-  }
-
-  return info->is_excluded();
-}
-
 // Check if a class or any of its supertypes has been redefined.
 bool SystemDictionaryShared::has_been_redefined(InstanceKlass* k) {
   if (k->has_been_redefined()) {
@@ -1401,94 +586,11 @@ bool SystemDictionaryShared::has_been_redefined(InstanceKlass* k) {
   return false;
 }
 
-bool SystemDictionaryShared::check_for_exclusion_impl(InstanceKlass* k) {
-  if (k->is_in_error_state()) {
-    return warn_excluded(k, "In error state");
-  }
-  if (has_been_redefined(k)) {
-    return warn_excluded(k, "Has been redefined");
-  }
-  if (!k->is_hidden() && k->shared_classpath_index() < 0 && is_builtin(k)) {
-    // These are classes loaded from unsupported locations (such as those loaded by JVMTI native
-    // agent during dump time).
-    return warn_excluded(k, "Unsupported location");
-  }
-  if (k->signers() != NULL) {
-    // We cannot include signed classes in the archive because the certificates
-    // used during dump time may be different than those used during
-    // runtime (due to expiration, etc).
-    return warn_excluded(k, "Signed JAR");
-  }
-  if (is_jfr_event_class(k)) {
-    // We cannot include JFR event classes because they need runtime-specific
-    // instrumentation in order to work with -XX:FlightRecorderOptions:retransform=false.
-    // There are only a small number of these classes, so it's not worthwhile to
-    // support them and make CDS more complicated.
-    return warn_excluded(k, "JFR event class");
-  }
-  if (k->init_state() < InstanceKlass::linked) {
-    // In CDS dumping, we will attempt to link all classes. Those that fail to link will
-    // be recorded in DumpTimeSharedClassInfo.
-    Arguments::assert_is_dumping_archive();
-
-    // TODO -- rethink how this can be handled.
-    // We should try to link ik, however, we can't do it here because
-    // 1. We are at VM exit
-    // 2. linking a class may cause other classes to be loaded, which means
-    //    a custom ClassLoader.loadClass() may be called, at a point where the
-    //    class loader doesn't expect it.
-    if (has_class_failed_verification(k)) {
-      return warn_excluded(k, "Failed verification");
-    } else {
-      if (k->can_be_verified_at_dumptime()) {
-        return warn_excluded(k, "Not linked");
-      }
-    }
-  }
-  if (DynamicDumpSharedSpaces && k->major_version() < 50 /*JAVA_6_VERSION*/) {
-    // In order to support old classes during dynamic dump, class rewriting needs to
-    // be reverted. This would result in more complex code and testing but not much gain.
-    ResourceMark rm;
-    log_warning(cds)("Pre JDK 6 class not supported by CDS: %u.%u %s",
-                     k->major_version(),  k->minor_version(), k->name()->as_C_string());
-    return true;
-  }
-
-  if (!k->can_be_verified_at_dumptime() && k->is_linked()) {
-    return warn_excluded(k, "Old class has been linked");
-  }
-
-  if (k->is_hidden() && !is_registered_lambda_proxy_class(k)) {
-    ResourceMark rm;
-    log_debug(cds)("Skipping %s: Hidden class", k->name()->as_C_string());
-    return true;
-  }
-
-  InstanceKlass* super = k->java_super();
-  if (super != NULL && check_for_exclusion(super, NULL)) {
-    ResourceMark rm;
-    log_warning(cds)("Skipping %s: super class %s is excluded", k->name()->as_C_string(), super->name()->as_C_string());
-    return true;
-  }
-
-  Array<InstanceKlass*>* interfaces = k->local_interfaces();
-  int len = interfaces->length();
-  for (int i = 0; i < len; i++) {
-    InstanceKlass* intf = interfaces->at(i);
-    if (check_for_exclusion(intf, NULL)) {
-      log_warning(cds)("Skipping %s: interface %s is excluded", k->name()->as_C_string(), intf->name()->as_C_string());
-      return true;
-    }
-  }
-
-  return false; // false == k should NOT be excluded
-}
-
 // k is a class before relocating by ArchiveBuilder
 void SystemDictionaryShared::validate_before_archiving(InstanceKlass* k) {
   ResourceMark rm;
   const char* name = k->name()->as_C_string();
-  DumpTimeSharedClassInfo* info = _dumptime_table->get(k);
+  DumpTimeClassInfo* info = _dumptime_table->get(k);
   assert(_no_class_loading_should_happen, "class loading must be disabled");
   guarantee(info != NULL, "Class %s must be entered into _dumptime_table", name);
   guarantee(!info->is_excluded(), "Should not attempt to archive excluded class %s", name);
@@ -1511,7 +613,7 @@ class UnregisteredClassesDuplicationChecker : StackObj {
 public:
   UnregisteredClassesDuplicationChecker() : _thread(Thread::current()) {}
 
-  bool do_entry(InstanceKlass* k, DumpTimeSharedClassInfo& info) {
+  bool do_entry(InstanceKlass* k, DumpTimeClassInfo& info) {
     if (!SystemDictionaryShared::is_builtin(k)) {
       _list.append(k);
     }
@@ -1549,7 +651,7 @@ public:
 
 class ExcludeDumpTimeSharedClasses : StackObj {
 public:
-  bool do_entry(InstanceKlass* k, DumpTimeSharedClassInfo& info) {
+  bool do_entry(InstanceKlass* k, DumpTimeClassInfo& info) {
     SystemDictionaryShared::check_for_exclusion(k, &info);
     return true; // keep on iterating
   }
@@ -1577,14 +679,14 @@ bool SystemDictionaryShared::is_excluded_class(InstanceKlass* k) {
   assert(_no_class_loading_should_happen, "sanity");
   assert_lock_strong(DumpTimeTable_lock);
   Arguments::assert_is_dumping_archive();
-  DumpTimeSharedClassInfo* p = find_or_allocate_info_for_locked(k);
+  DumpTimeClassInfo* p = find_or_allocate_info_for_locked(k);
   return (p == NULL) ? true : p->is_excluded();
 }
 
 void SystemDictionaryShared::set_excluded_locked(InstanceKlass* k) {
   assert_lock_strong(DumpTimeTable_lock);
   Arguments::assert_is_dumping_archive();
-  DumpTimeSharedClassInfo* info = find_or_allocate_info_for_locked(k);
+  DumpTimeClassInfo* info = find_or_allocate_info_for_locked(k);
   if (info != NULL) {
     info->set_excluded();
   }
@@ -1592,7 +694,7 @@ void SystemDictionaryShared::set_excluded_locked(InstanceKlass* k) {
 
 void SystemDictionaryShared::set_excluded(InstanceKlass* k) {
   Arguments::assert_is_dumping_archive();
-  DumpTimeSharedClassInfo* info = find_or_allocate_info_for(k);
+  DumpTimeClassInfo* info = find_or_allocate_info_for(k);
   if (info != NULL) {
     info->set_excluded();
   }
@@ -1600,7 +702,7 @@ void SystemDictionaryShared::set_excluded(InstanceKlass* k) {
 
 void SystemDictionaryShared::set_class_has_failed_verification(InstanceKlass* ik) {
   Arguments::assert_is_dumping_archive();
-  DumpTimeSharedClassInfo* p = find_or_allocate_info_for(ik);
+  DumpTimeClassInfo* p = find_or_allocate_info_for(ik);
   if (p != NULL) {
     p->set_failed_verification();
   }
@@ -1613,7 +715,7 @@ bool SystemDictionaryShared::has_class_failed_verification(InstanceKlass* ik) {
     assert(ik->is_shared(), "must be a shared class in the static archive");
     return false;
   }
-  DumpTimeSharedClassInfo* p = _dumptime_table->get(ik);
+  DumpTimeClassInfo* p = _dumptime_table->get(ik);
   return (p == NULL) ? false : p->failed_verification();
 }
 
@@ -1622,7 +724,7 @@ class IterateDumpTimeSharedClassTable : StackObj {
 public:
   IterateDumpTimeSharedClassTable(MetaspaceClosure* it) : _it(it) {}
 
-  bool do_entry(InstanceKlass* k, DumpTimeSharedClassInfo& info) {
+  bool do_entry(InstanceKlass* k, DumpTimeClassInfo& info) {
     assert_lock_strong(DumpTimeTable_lock);
     if (k->is_loader_alive() && !info.is_excluded()) {
       info.metaspace_pointers_do(_it);
@@ -1659,7 +761,7 @@ void SystemDictionaryShared::dumptime_classes_do(class MetaspaceClosure* it) {
 bool SystemDictionaryShared::add_verification_constraint(InstanceKlass* k, Symbol* name,
          Symbol* from_name, bool from_field_is_protected, bool from_is_array, bool from_is_object) {
   Arguments::assert_is_dumping_archive();
-  DumpTimeSharedClassInfo* info = find_or_allocate_info_for(k);
+  DumpTimeClassInfo* info = find_or_allocate_info_for(k);
   if (info != NULL) {
     info->add_verification_constraint(k, name, from_name, from_field_is_protected,
                                       from_is_array, from_is_object);
@@ -1685,36 +787,23 @@ bool SystemDictionaryShared::add_verification_constraint(InstanceKlass* k, Symbo
   }
 }
 
-void DumpTimeSharedClassInfo::add_verification_constraint(InstanceKlass* k, Symbol* name,
-         Symbol* from_name, bool from_field_is_protected, bool from_is_array, bool from_is_object) {
-  if (_verifier_constraints == NULL) {
-    _verifier_constraints = new(ResourceObj::C_HEAP, mtClass) GrowableArray<DTVerifierConstraint>(4, mtClass);
+void SystemDictionaryShared::add_to_dump_time_lambda_proxy_class_dictionary(LambdaProxyClassKey& key,
+                                                           InstanceKlass* proxy_klass) {
+  assert_lock_strong(DumpTimeTable_lock);
+  if (_dumptime_lambda_proxy_class_dictionary == NULL) {
+    _dumptime_lambda_proxy_class_dictionary =
+      new (ResourceObj::C_HEAP, mtClass)DumpTimeLambdaProxyClassDictionary();
   }
-  if (_verifier_constraint_flags == NULL) {
-    _verifier_constraint_flags = new(ResourceObj::C_HEAP, mtClass) GrowableArray<char>(4, mtClass);
-  }
-  GrowableArray<DTVerifierConstraint>* vc_array = _verifier_constraints;
-  for (int i = 0; i < vc_array->length(); i++) {
-    DTVerifierConstraint* p = vc_array->adr_at(i);
-    if (name == p->_name && from_name == p->_from_name) {
-      return;
-    }
-  }
-  DTVerifierConstraint cons(name, from_name);
-  vc_array->append(cons);
-
-  GrowableArray<char>* vcflags_array = _verifier_constraint_flags;
-  char c = 0;
-  c |= from_field_is_protected ? SystemDictionaryShared::FROM_FIELD_IS_PROTECTED : 0;
-  c |= from_is_array           ? SystemDictionaryShared::FROM_IS_ARRAY           : 0;
-  c |= from_is_object          ? SystemDictionaryShared::FROM_IS_OBJECT          : 0;
-  vcflags_array->append(c);
-
-  if (log_is_enabled(Trace, cds, verification)) {
-    ResourceMark rm;
-    log_trace(cds, verification)("add_verification_constraint: %s: %s must be subclass of %s [0x%x] array len %d flags len %d",
-                                 k->external_name(), from_name->as_klass_external_name(),
-                                 name->as_klass_external_name(), c, vc_array->length(), vcflags_array->length());
+  DumpTimeLambdaProxyClassInfo* lambda_info = _dumptime_lambda_proxy_class_dictionary->get(key);
+  if (lambda_info == NULL) {
+    DumpTimeLambdaProxyClassInfo info;
+    info.add_proxy_klass(proxy_klass);
+    _dumptime_lambda_proxy_class_dictionary->put(key, info);
+    //lambda_info = _dumptime_lambda_proxy_class_dictionary->get(key);
+    //assert(lambda_info->_proxy_klass == proxy_klass, "must be"); // debug only -- remove
+    ++_dumptime_lambda_proxy_class_dictionary->_count;
+  } else {
+    lambda_info->add_proxy_klass(proxy_klass);
   }
 }
 
@@ -1738,11 +827,11 @@ void SystemDictionaryShared::add_lambda_proxy_class(InstanceKlass* caller_ik,
   InstanceKlass* nest_host = caller_ik->nest_host(CHECK);
   assert(nest_host != NULL, "unexpected NULL nest_host");
 
-  DumpTimeSharedClassInfo* info = _dumptime_table->get(lambda_ik);
+  DumpTimeClassInfo* info = _dumptime_table->get(lambda_ik);
   if (info != NULL && !lambda_ik->is_non_strong_hidden() && is_builtin(lambda_ik) && is_builtin(caller_ik)
       // Don't include the lambda proxy if its nest host is not in the "linked" state.
       && nest_host->is_linked()) {
-    // Set _is_archived_lambda_proxy in DumpTimeSharedClassInfo so that the lambda_ik
+    // Set _is_archived_lambda_proxy in DumpTimeClassInfo so that the lambda_ik
     // won't be excluded during dumping of shared archive. See ExcludeDumpTimeSharedClasses.
     info->_is_archived_lambda_proxy = true;
     info->set_nest_host(nest_host);
@@ -1803,7 +892,7 @@ InstanceKlass* SystemDictionaryShared::get_shared_lambda_proxy_class(InstanceKla
 
 InstanceKlass* SystemDictionaryShared::get_shared_nest_host(InstanceKlass* lambda_ik) {
   assert(!DumpSharedSpaces && UseSharedSpaces, "called at run time with CDS enabled only");
-  RunTimeSharedClassInfo* record = RunTimeSharedClassInfo::get_for(lambda_ik);
+  RunTimeClassInfo* record = RunTimeClassInfo::get_for(lambda_ik);
   return record->nest_host();
 }
 
@@ -1813,7 +902,7 @@ InstanceKlass* SystemDictionaryShared::prepare_shared_lambda_proxy_class(Instanc
   Handle protection_domain;
   PackageEntry* pkg_entry = caller_ik->package();
   if (caller_ik->class_loader() != NULL) {
-    protection_domain = SystemDictionaryShared::init_security_info(class_loader, caller_ik, pkg_entry, CHECK_NULL);
+    protection_domain = CDSProtectionDomain::init_security_info(class_loader, caller_ik, pkg_entry, CHECK_NULL);
   }
 
   InstanceKlass* shared_nest_host = get_shared_nest_host(lambda_ik);
@@ -1853,73 +942,15 @@ InstanceKlass* SystemDictionaryShared::prepare_shared_lambda_proxy_class(Instanc
   return loaded_lambda;
 }
 
-static char get_loader_type_by(oop  loader) {
-  assert(SystemDictionary::is_builtin_class_loader(loader), "Must be built-in loader");
-  if (SystemDictionary::is_boot_class_loader(loader)) {
-    return (char)ClassLoader::BOOT_LOADER;
-  } else if (SystemDictionary::is_platform_class_loader(loader)) {
-    return (char)ClassLoader::PLATFORM_LOADER;
-  } else {
-    assert(SystemDictionary::is_system_class_loader(loader), "Class loader mismatch");
-    return (char)ClassLoader::APP_LOADER;
-  }
-}
-
-static oop get_class_loader_by(char type) {
-  if (type == (char)ClassLoader::BOOT_LOADER) {
-    return (oop)NULL;
-  } else if (type == (char)ClassLoader::PLATFORM_LOADER) {
-    return SystemDictionary::java_platform_loader();
-  } else {
-    assert (type == (char)ClassLoader::APP_LOADER, "Sanity");
-    return SystemDictionary::java_system_loader();
-  }
-}
-
-void DumpTimeSharedClassInfo::record_linking_constraint(Symbol* name, Handle loader1, Handle loader2) {
-  assert(loader1 != loader2, "sanity");
-  LogTarget(Info, class, loader, constraints) log;
-  if (_loader_constraints == NULL) {
-    _loader_constraints = new (ResourceObj::C_HEAP, mtClass) GrowableArray<DTLoaderConstraint>(4, mtClass);
-  }
-  char lt1 = get_loader_type_by(loader1());
-  char lt2 = get_loader_type_by(loader2());
-  DTLoaderConstraint lc(name, lt1, lt2);
-  for (int i = 0; i < _loader_constraints->length(); i++) {
-    DTLoaderConstraint dt = _loader_constraints->at(i);
-    if (lc.equals(dt)) {
-      if (log.is_enabled()) {
-        ResourceMark rm;
-        // Use loader[0]/loader[1] to be consistent with the logs in loaderConstraints.cpp
-        log.print("[CDS record loader constraint for class: %s constraint_name: %s loader[0]: %s loader[1]: %s already added]",
-                  _klass->external_name(), name->as_C_string(),
-                  ClassLoaderData::class_loader_data(loader1())->loader_name_and_id(),
-                  ClassLoaderData::class_loader_data(loader2())->loader_name_and_id());
-      }
-      return;
-    }
-  }
-  _loader_constraints->append(lc);
-  if (log.is_enabled()) {
-    ResourceMark rm;
-    // Use loader[0]/loader[1] to be consistent with the logs in loaderConstraints.cpp
-    log.print("[CDS record loader constraint for class: %s constraint_name: %s loader[0]: %s loader[1]: %s total %d]",
-              _klass->external_name(), name->as_C_string(),
-              ClassLoaderData::class_loader_data(loader1())->loader_name_and_id(),
-              ClassLoaderData::class_loader_data(loader2())->loader_name_and_id(),
-              _loader_constraints->length());
-  }
-}
-
 void SystemDictionaryShared::check_verification_constraints(InstanceKlass* klass,
                                                             TRAPS) {
   assert(!DumpSharedSpaces && UseSharedSpaces, "called at run time with CDS enabled only");
-  RunTimeSharedClassInfo* record = RunTimeSharedClassInfo::get_for(klass);
+  RunTimeClassInfo* record = RunTimeClassInfo::get_for(klass);
 
   int length = record->_num_verifier_constraints;
   if (length > 0) {
     for (int i = 0; i < length; i++) {
-      RunTimeSharedClassInfo::RTVerifierConstraint* vc = record->verifier_constraint_at(i);
+      RunTimeClassInfo::RTVerifierConstraint* vc = record->verifier_constraint_at(i);
       Symbol* name      = vc->name();
       Symbol* from_name = vc->from_name();
       char c            = record->verifier_constraint_flag(i);
@@ -1949,6 +980,17 @@ void SystemDictionaryShared::check_verification_constraints(InstanceKlass* klass
         THROW_MSG(vmSymbols::java_lang_VerifyError(), ss.as_string());
       }
     }
+  }
+}
+
+static oop get_class_loader_by(char type) {
+  if (type == (char)ClassLoader::BOOT_LOADER) {
+    return (oop)NULL;
+  } else if (type == (char)ClassLoader::PLATFORM_LOADER) {
+    return SystemDictionary::java_platform_loader();
+  } else {
+    assert (type == (char)ClassLoader::APP_LOADER, "Sanity");
+    return SystemDictionary::java_system_loader();
   }
 }
 
@@ -2012,7 +1054,7 @@ void SystemDictionaryShared::record_linking_constraint(Symbol* name, InstanceKla
   assert(!Thread::current()->is_VM_thread(), "must be");
 
   Arguments::assert_is_dumping_archive();
-  DumpTimeSharedClassInfo* info = find_or_allocate_info_for(klass);
+  DumpTimeClassInfo* info = find_or_allocate_info_for(klass);
   if (info != NULL) {
     info->record_linking_constraint(name, loader1, loader2);
   }
@@ -2028,12 +1070,12 @@ bool SystemDictionaryShared::check_linking_constraints(Thread* current, Instance
     return true;
   }
   if (klass->is_shared_platform_class() || klass->is_shared_app_class()) {
-    RunTimeSharedClassInfo* info = RunTimeSharedClassInfo::get_for(klass);
+    RunTimeClassInfo* info = RunTimeClassInfo::get_for(klass);
     assert(info != NULL, "Sanity");
     if (info->_num_loader_constraints > 0) {
       HandleMark hm(current);
       for (int i = 0; i < info->_num_loader_constraints; i++) {
-        RunTimeSharedClassInfo::RTLoaderConstraint* lc = info->loader_constraint_at(i);
+        RunTimeClassInfo::RTLoaderConstraint* lc = info->loader_constraint_at(i);
         Symbol* name = lc->constraint_name();
         Handle loader1(current, get_class_loader_by(lc->_loader_type1));
         Handle loader2(current, get_class_loader_by(lc->_loader_type2));
@@ -2124,9 +1166,9 @@ public:
     _num_unregistered_klasses = 0;
   }
 
-  bool do_entry(InstanceKlass* k, DumpTimeSharedClassInfo& info) {
+  bool do_entry(InstanceKlass* k, DumpTimeClassInfo& info) {
     if (!info.is_excluded()) {
-      size_t byte_size = RunTimeSharedClassInfo::byte_size(info._klass, info.num_verifier_constraints(), info.num_loader_constraints());
+      size_t byte_size = RunTimeClassInfo::byte_size(info._klass, info.num_verifier_constraints(), info.num_loader_constraints());
       _shared_class_info_size += align_up(byte_size, SharedSpaceObjectAlignment);
     }
     return true; // keep on iterating
@@ -2233,11 +1275,11 @@ public:
                                bool is_builtin)
     : _writer(writer), _is_builtin(is_builtin), _builder(ArchiveBuilder::current()) {}
 
-  bool do_entry(InstanceKlass* k, DumpTimeSharedClassInfo& info) {
+  bool do_entry(InstanceKlass* k, DumpTimeClassInfo& info) {
     if (!info.is_excluded() && info.is_builtin() == _is_builtin) {
-      size_t byte_size = RunTimeSharedClassInfo::byte_size(info._klass, info.num_verifier_constraints(), info.num_loader_constraints());
-      RunTimeSharedClassInfo* record;
-      record = (RunTimeSharedClassInfo*)ArchiveBuilder::ro_region_alloc(byte_size);
+      size_t byte_size = RunTimeClassInfo::byte_size(info._klass, info.num_verifier_constraints(), info.num_loader_constraints());
+      RunTimeClassInfo* record;
+      record = (RunTimeClassInfo*)ArchiveBuilder::ro_region_alloc(byte_size);
       record->init(info);
 
       unsigned int hash;
@@ -2254,8 +1296,8 @@ public:
         log_trace(cds,hashtables)("%s dictionary: %s", (_is_builtin ? "builtin" : "unregistered"), info._klass->external_name());
       }
 
-      // Save this for quick runtime lookup of InstanceKlass* -> RunTimeSharedClassInfo*
-      RunTimeSharedClassInfo::set_for(info._klass, record);
+      // Save this for quick runtime lookup of InstanceKlass* -> RunTimeClassInfo*
+      RunTimeClassInfo::set_for(info._klass, record);
     }
     return true; // keep on iterating
   }
@@ -2327,7 +1369,7 @@ void SystemDictionaryShared::serialize_vm_classes(SerializeClosure* soc) {
   }
 }
 
-const RunTimeSharedClassInfo*
+const RunTimeClassInfo*
 SystemDictionaryShared::find_record(RunTimeSharedDictionary* static_dict, RunTimeSharedDictionary* dynamic_dict, Symbol* name) {
   if (!UseSharedSpaces || !name->is_shared()) {
     // The names of all shared classes must also be a shared Symbol.
@@ -2335,7 +1377,7 @@ SystemDictionaryShared::find_record(RunTimeSharedDictionary* static_dict, RunTim
   }
 
   unsigned int hash = SystemDictionaryShared::hash_for_shared_dictionary_quick(name);
-  const RunTimeSharedClassInfo* record = NULL;
+  const RunTimeClassInfo* record = NULL;
   if (DynamicArchive::is_mapped()) {
     // Those regenerated holder classes are in dynamic archive
     if (name == vmSymbols::java_lang_invoke_Invokers_Holder() ||
@@ -2363,7 +1405,7 @@ SystemDictionaryShared::find_record(RunTimeSharedDictionary* static_dict, RunTim
 }
 
 InstanceKlass* SystemDictionaryShared::find_builtin_class(Symbol* name) {
-  const RunTimeSharedClassInfo* record = find_record(&_builtin_dictionary, &_dynamic_builtin_dictionary, name);
+  const RunTimeClassInfo* record = find_record(&_builtin_dictionary, &_dynamic_builtin_dictionary, name);
   if (record != NULL) {
     assert(!record->_klass->is_hidden(), "hidden class cannot be looked up by name");
     assert(check_alignment(record->_klass), "Address not aligned");
@@ -2375,7 +1417,7 @@ InstanceKlass* SystemDictionaryShared::find_builtin_class(Symbol* name) {
 
 void SystemDictionaryShared::update_shared_entry(InstanceKlass* k, int id) {
   assert(DumpSharedSpaces, "supported only when dumping");
-  DumpTimeSharedClassInfo* info = find_or_allocate_info_for(k);
+  DumpTimeClassInfo* info = find_or_allocate_info_for(k);
   info->_id = id;
 }
 
@@ -2403,7 +1445,7 @@ class SharedDictionaryPrinter : StackObj {
 public:
   SharedDictionaryPrinter(outputStream* st) : _st(st), _index(0) {}
 
-  void do_value(const RunTimeSharedClassInfo* record) {
+  void do_value(const RunTimeClassInfo* record) {
     ResourceMark rm;
     _st->print_cr("%4d: %s %s", _index++, record->_klass->external_name(),
         class_loader_name_for_shared(record->_klass));
@@ -2484,7 +1526,7 @@ void SystemDictionaryShared::print_table_statistics(outputStream* st) {
   }
 }
 
-bool SystemDictionaryShared::empty_dumptime_table() {
+bool SystemDictionaryShared::is_dumptime_table_empty() {
   if (_dumptime_table == NULL) {
     return true;
   }
@@ -2516,7 +1558,7 @@ public:
     }
   }
 
-  void do_value(const RunTimeSharedClassInfo* info) {
+  void do_value(const RunTimeClassInfo* info) {
     InstanceKlass* ik = info->_klass;
     update(ik);
     update_array_klasses(ik->array_klasses());
